+++ zcash_client_backend/src/data_api.rs
@@ -1447,7 +1447,7 @@ pub trait InputSource {
         &self,
         _outpoint: &OutPoint,
     ) -> Result<Option<WalletTransparentOutput>, Self::Error> {
-        Ok(None)
+        unimplemented!("InputSource::get_spendable_transparent_output must be overridden for wallets to use the `transparent-inputs` feature")
     }
 
     /// Returns the list of spendable transparent outputs received by this wallet at `address`
@@ -1465,7 +1465,7 @@ pub trait InputSource {
         _target_height: TargetHeight,
         _confirmations_policy: ConfirmationsPolicy,
     ) -> Result<Vec<WalletTransparentOutput>, Self::Error> {
-        Ok(vec![])
+        unimplemented!("InputSource::get_spendable_transparent_outputs must be overridden for wallets to use the `transparent-inputs` feature")
     }
 }
 
@@ -1665,8 +1665,10 @@ pub trait WalletRead {
     #[cfg(feature = "orchard")]
     fn get_orchard_nullifiers(
         &self,
-        query: NullifierQuery,
-    ) -> Result<Vec<(Self::AccountId, orchard::note::Nullifier)>, Self::Error>;
+        _query: NullifierQuery,
+    ) -> Result<Vec<(Self::AccountId, orchard::note::Nullifier)>, Self::Error> {
+        unimplemented!("WalletRead::get_orchard_nullifiers must be overridden for wallets to use the `orchard` feature")
+    }
 
     /// Returns the set of non-ephemeral transparent receivers associated with the given
     /// account controlled by this wallet.
@@ -1693,7 +1695,7 @@ pub trait WalletRead {
         _include_change: bool,
         _include_standalone: bool,
     ) -> Result<HashMap<TransparentAddress, Option<TransparentAddressMetadata>>, Self::Error> {
-        Ok(HashMap::new())
+        unimplemented!("WalletRead::get_transparent_receivers must be overridden for wallets to use the `transparent-inputs` feature")
     }
 
     /// Returns a mapping from each transparent receiver associated with the specified account
@@ -1707,7 +1709,7 @@ pub trait WalletRead {
         _target_height: TargetHeight,
         _confirmations_policy: ConfirmationsPolicy,
     ) -> Result<HashMap<TransparentAddress, Balance>, Self::Error> {
-        Ok(HashMap::new())
+        unimplemented!("WalletRead::get_transparent_balances must be overridden for wallets to use the `transparent-inputs` feature")
     }
 
     /// Returns the metadata associated with a given transparent receiver in an account
@@ -1761,7 +1763,9 @@ pub trait WalletRead {
     /// received an output in a mined transaction; therefore, we take that height to be where we
     /// should start searching for UTXOs.
     #[cfg(feature = "transparent-inputs")]
-    fn utxo_query_height(&self, account: Self::AccountId) -> Result<BlockHeight, Self::Error>;
+    fn utxo_query_height(&self, _account: Self::AccountId) -> Result<BlockHeight, Self::Error> {
+        unimplemented!("WalletRead::utxo_query_height must be overridden for wallets to use the `transparent-inputs` feature")
+    }
 
     /// Returns a vector of ephemeral transparent addresses associated with the given
     /// account controlled by this wallet, along with their metadata. The result includes
@@ -1802,7 +1806,7 @@ pub trait WalletRead {
         _account: Self::AccountId,
         _index_range: Option<Range<NonHardenedChildIndex>>,
     ) -> Result<Vec<(TransparentAddress, TransparentAddressMetadata)>, Self::Error> {
-        Ok(vec![])
+        unimplemented!("WalletRead::get_known_ephemeral_addresses must be overridden for wallets to use the `transparent-inputs` feature")
     }
 
     /// If a given ephemeral address might have been reserved, i.e. would be included in
@@ -1905,9 +1909,11 @@ pub trait WalletTest: InputSource + WalletRead {
     #[cfg(feature = "transparent-inputs")]
     fn get_transparent_output(
         &self,
-        outpoint: &OutPoint,
-        allow_unspendable: bool,
-    ) -> Result<Option<WalletTransparentOutput>, <Self as InputSource>::Error>;
+        _outpoint: &OutPoint,
+        _allow_unspendable: bool,
+    ) -> Result<Option<WalletTransparentOutput>, <Self as InputSource>::Error> {
+        unimplemented!("WalletTest::get_transparent_output must be overridden for wallets to use the `transparent-inputs` feature")
+    }
 
     /// Returns all the notes that have been received by the wallet.
     fn get_notes(
@@ -2760,9 +2766,11 @@ pub trait WalletWrite: WalletRead {
     #[cfg(feature = "transparent-key-import")]
     fn import_standalone_transparent_pubkey(
         &mut self,
-        account: Self::AccountId,
-        pubkey: secp256k1::PublicKey,
-    ) -> Result<(), Self::Error>;
+        _account: Self::AccountId,
+        _pubkey: secp256k1::PublicKey,
+    ) -> Result<(), Self::Error> {
+        unimplemented!("WalletWrite::import_standalone_transparent_pubkey must be overridden for wallets to use the `transparent-key-import` feature")
+    }
 
     /// Generates, persists, and marks as exposed the next available diversified address for the
     /// specified account, given the current addresses known to the wallet.
@@ -2893,10 +2901,7 @@ pub trait WalletWrite: WalletRead {
         _account_id: Self::AccountId,
         _n: usize,
     ) -> Result<Vec<(TransparentAddress, TransparentAddressMetadata)>, Self::Error> {
-        // Default impl is required for feature-flagged trait methods to prevent
-        // breakage due to inadvertent activation of features by transitive dependencies
-        // of the implementing crate.
-        Ok(vec![])
+        unimplemented!("WalletWrite::reserve_next_n_ephemeral_addresses must be overridden for wallets to use the `transparent-inputs` feature")
     }
 
     /// Updates the wallet backend with respect to the status of a specific transaction, from the
@@ -2922,9 +2927,11 @@ pub trait WalletWrite: WalletRead {
     #[cfg(feature = "transparent-inputs")]
     fn notify_address_checked(
         &mut self,
-        request: TransactionsInvolvingAddress,
-        as_of_height: BlockHeight,
-    ) -> Result<(), Self::Error>;
+        _request: TransactionsInvolvingAddress,
+        _as_of_height: BlockHeight,
+    ) -> Result<(), Self::Error> {
+        unimplemented!("WalletWrite::notify_address_checked must be overridden for wallets to use the `transparent-inputs` feature")
+    }
 }
 
 /// This trait describes a capability for manipulating wallet note commitment trees.